Refactor combine qualifier traits
Move `add_const_from` and `add_reference_from` to `fatal/type/qualifier.h`.

It will be a bit easier to deal with these when adding more variants
later.
