+++ fatal/type/qualifier.h
@@ -10,8 +10,6 @@
 #ifndef FATAL_INCLUDE_fatal_type_qualifier_h
 #define FATAL_INCLUDE_fatal_type_qualifier_h
 
-#include <fatal/type/deprecated/type_list.h>
-
 #include <type_traits>
 
 namespace fatal {
@@ -61,6 +59,79 @@ enum class ref_qualifier:
   none = 0,
   lvalue = 1,
   rvalue = 2
+};
+
+/**
+ * Applies `std::add_const` to a type iff some other type is const.
+ *
+ * Example:
+ *
+ *  struct foo {};
+ *
+ *  // yields `foo`
+ *  using result1 = add_const_from<foo, int>::type;
+ *
+ *  // yields `foo const`
+ *  using result2 = add_const_from<foo, int const>::type;
+ *
+ *  // yields `foo const`
+ *  using result3 = add_const_from<foo const, int const>::type;
+ *
+ * @author: Marcelo Juchem <marcelo@fb.com>
+ */
+template <typename T, typename>
+struct add_const_from {
+  using type = T;
+};
+
+template <typename T, typename TFrom>
+struct add_const_from<T, TFrom const> {
+  using type = typename std::add_const<T>::type;
+};
+
+/**
+ * Given types `T` and `U`:
+ * - if `U` is not a reference, yield `T`
+ * - if `U` is an l-value reference, yield `std::add_lvalue_reference<T>::type`
+ * - if `U` is an r-value reference, yield `std::add_rvalue_reference<T>::type`
+ *
+ * Example:
+ *
+ *  struct foo {};
+ *
+ *  // yields `foo`
+ *  using result1 = add_reference_from<foo, int>::type;
+ *
+ *  // yields `foo &&`
+ *  using result2 = add_reference_from<foo &&, int>::type;
+ *
+ *  // yields `foo &`
+ *  using result3 = add_reference_from<foo, int &>::type;
+ *
+ *  // yields `foo &`
+ *  using result4 = add_reference_from<foo &&, int &>::type;
+ *
+ *  // yields `foo &&`
+ *  using result5 = add_reference_from<foo, int &&>::type;
+ *
+ *  // yields `foo &`
+ *  using result6 = add_reference_from<foo &, int &&>::type;
+ *
+ * @author: Marcelo Juchem <marcelo@fb.com>
+ */
+template <typename T, typename>
+struct add_reference_from {
+  using type = T;
+};
+
+template <typename T, typename TFrom>
+struct add_reference_from<T, TFrom &> {
+  using type = typename std::add_lvalue_reference<T>::type;
+};
+
+template <typename T, typename TFrom>
+struct add_reference_from<T, TFrom &&> {
+  using type = typename std::add_rvalue_reference<T>::type;
 };
 
 } // namespace fatal {
