+++ fatal/type/test/qualifier_test.cpp
@@ -21,7 +21,7 @@ struct foo {
   void vc() volatile const;
 };
 
-FATAL_TEST(reflect_member_function, cv_qualifier_bitwise_and) {
+FATAL_TEST(qualifier, cv_qualifier_bitwise_and) {
   FATAL_EXPECT_FALSE(cv_qualifier::none & cv_qualifier::none);
   FATAL_EXPECT_FALSE(cv_qualifier::none & cv_qualifier::c);
   FATAL_EXPECT_FALSE(cv_qualifier::none & cv_qualifier::v);
@@ -41,6 +41,115 @@ FATAL_TEST(reflect_member_function, cv_qualifier_bitwise_and) {
   FATAL_EXPECT_TRUE(cv_qualifier::cv & cv_qualifier::c);
   FATAL_EXPECT_TRUE(cv_qualifier::cv & cv_qualifier::v);
   FATAL_EXPECT_TRUE(cv_qualifier::cv & cv_qualifier::cv);
+}
+
+FATAL_TEST(qualifier, add_const_from) {
+# define TEST_IMPL(From, T, ...) \
+  FATAL_EXPECT_SAME<__VA_ARGS__, add_const_from<T, From>::type>();
+
+  TEST_IMPL(int, int &&, int &&);
+  TEST_IMPL(int, int &, int &);
+  TEST_IMPL(int, int *&&, int *&&);
+  TEST_IMPL(int, int *&, int *&);
+  TEST_IMPL(int, int *, int *);
+  TEST_IMPL(int, int, int);
+  TEST_IMPL(int, int *const &&, int *const &&);
+  TEST_IMPL(int, int *const &, int *const &);
+  TEST_IMPL(int, int *const, int *const);
+  TEST_IMPL(int, int const &&, int const &&);
+  TEST_IMPL(int, int const &, int const &);
+  TEST_IMPL(int, int const *&&, int const *&&);
+  TEST_IMPL(int, int const *&, int const *&);
+  TEST_IMPL(int, int const *, int const *);
+  TEST_IMPL(int, int const *const &&, int const *const &&);
+  TEST_IMPL(int, int const *const &, int const *const &);
+  TEST_IMPL(int, int const *const, int const *const);
+  TEST_IMPL(int, int const, int const);
+
+  TEST_IMPL(int const, int &&, int &&);
+  TEST_IMPL(int const, int &, int &);
+  TEST_IMPL(int const, int, int const);
+  TEST_IMPL(int const, int *&&, int *&&);
+  TEST_IMPL(int const, int *&, int *&);
+  TEST_IMPL(int const, int *, int *const);
+  TEST_IMPL(int const, int *const &&, int *const &&);
+  TEST_IMPL(int const, int *const &, int *const &);
+  TEST_IMPL(int const, int *const, int *const);
+  TEST_IMPL(int const, int const &&, int const &&);
+  TEST_IMPL(int const, int const &, int const &);
+  TEST_IMPL(int const, int const, int const);
+  TEST_IMPL(int const, int const *&&, int const *&&);
+  TEST_IMPL(int const, int const *&, int const *&);
+  TEST_IMPL(int const, int const *, int const *const);
+  TEST_IMPL(int const, int const *const &&, int const *const &&);
+  TEST_IMPL(int const, int const *const &, int const *const &);
+  TEST_IMPL(int const, int const *const, int const *const);
+
+# undef TEST_IMPL
+}
+
+FATAL_TEST(qualifier, add_reference_from) {
+# define TEST_IMPL(From, T, ...) \
+  FATAL_EXPECT_SAME<__VA_ARGS__, add_reference_from<T, From>::type>();
+
+  TEST_IMPL(int &&, int &&, int &&);
+  TEST_IMPL(int &&, int &, int &);
+  TEST_IMPL(int &&, int, int &&);
+  TEST_IMPL(int &&, int *&&, int *&&);
+  TEST_IMPL(int &&, int *&, int *&);
+  TEST_IMPL(int &&, int *, int * &&);
+  TEST_IMPL(int &&, int *const &&, int *const &&);
+  TEST_IMPL(int &&, int *const &, int *const &);
+  TEST_IMPL(int &&, int *const, int *const &&);
+  TEST_IMPL(int &&, int const &&, int const &&);
+  TEST_IMPL(int &&, int const &, int const &);
+  TEST_IMPL(int &&, int const, int const &&);
+  TEST_IMPL(int &&, int const *&&, int const *&&);
+  TEST_IMPL(int &&, int const *&, int const *&);
+  TEST_IMPL(int &&, int const *, int const * &&);
+  TEST_IMPL(int &&, int const *const &&, int const *const &&);
+  TEST_IMPL(int &&, int const *const &, int const *const &);
+  TEST_IMPL(int &&, int const *const, int const *const &&);
+
+  TEST_IMPL(int &, int &&, int &);
+  TEST_IMPL(int &, int &, int &);
+  TEST_IMPL(int &, int, int &);
+  TEST_IMPL(int &, int *&&, int *&);
+  TEST_IMPL(int &, int *&, int *&);
+  TEST_IMPL(int &, int *, int *&);
+  TEST_IMPL(int &, int *const &&, int *const &);
+  TEST_IMPL(int &, int *const &, int *const &);
+  TEST_IMPL(int &, int *const, int *const &);
+  TEST_IMPL(int &, int const &&, int const &);
+  TEST_IMPL(int &, int const &, int const &);
+  TEST_IMPL(int &, int const, int const &);
+  TEST_IMPL(int &, int const *&&, int const *&);
+  TEST_IMPL(int &, int const *&, int const *&);
+  TEST_IMPL(int &, int const *, int const *&);
+  TEST_IMPL(int &, int const *const &&, int const *const &);
+  TEST_IMPL(int &, int const *const &, int const *const &);
+  TEST_IMPL(int &, int const *const, int const *const &);
+
+  TEST_IMPL(int, int &&, int &&);
+  TEST_IMPL(int, int &, int &);
+  TEST_IMPL(int, int, int);
+  TEST_IMPL(int, int *&&, int *&&);
+  TEST_IMPL(int, int *&, int *&);
+  TEST_IMPL(int, int *, int *);
+  TEST_IMPL(int, int *const &&, int *const &&);
+  TEST_IMPL(int, int *const &, int *const &);
+  TEST_IMPL(int, int *const, int *const);
+  TEST_IMPL(int, int const &&, int const &&);
+  TEST_IMPL(int, int const &, int const &);
+  TEST_IMPL(int, int const, int const);
+  TEST_IMPL(int, int const *&&, int const *&&);
+  TEST_IMPL(int, int const *&, int const *&);
+  TEST_IMPL(int, int const *, int const *);
+  TEST_IMPL(int, int const *const &&, int const *const &&);
+  TEST_IMPL(int, int const *const &, int const *const &);
+  TEST_IMPL(int, int const *const, int const *const);
+
+# undef TEST_IMPL
 }
 
 } // namespace fatal {
